extract CustomTimeline to separate component
